Refactor recent snapshot function for reuse
