+++ internal/kube/ptr.go
@@ -1,0 +1,5 @@
+package kube
+
+func ptr[T any](v T) *T {
+	return &v
+}
